fix/ delete controller

+++ controllers/movieControllers.js
@@ -1,11 +1,8 @@
-const ForbiddenError = require('../errors/ForbiddenError');
 const NotFoundError = require('../errors/NotFoundError');
 const movie = require('../models/movie');
 const {
-  forbiddenForDeleteMovieTextError,
   successfulMovieDeleteText,
   notFoundMovieForDeleteTextError,
-  notFoundSavedMoviesTextError,
 } = require('../utils/constants');
 
 const getMovies = (req, res, next) => {
@@ -13,7 +10,6 @@ const getMovies = (req, res, next) => {
   movie
     .find({ owner: _id })
     .then((data) => {
-      if (!data[0]) next(new NotFoundError(notFoundSavedMoviesTextError));
       res.status(200).send(data);
     })
     .catch(next);
@@ -32,19 +28,10 @@ const createMovie = (req, res, next) => {
 const deleteMovie = (req, res, next) => {
   const { _id } = req.params;
   movie
-    .findOne({ movieId: _id })
-    .then((movieData) => {
-      if (movieData.owner.toString() !== req.user._id) {
-        return next(new ForbiddenError(forbiddenForDeleteMovieTextError));
-      }
-      return movie
-        .deleteOne({ movieId: _id })
-        .then(() => {
-          res.status(200).send({ message: successfulMovieDeleteText });
-        })
-        .catch(next);
+    .deleteOne({ movieId: _id, owner: req.user._id.toString() })
+    .then(() => {
+      res.status(200).send({ message: successfulMovieDeleteText });
     })
-
     .catch((err) => {
       if (err.name === 'TypeError') {
         return next(new NotFoundError(notFoundMovieForDeleteTextError));
